Make builds work on FreeBSD and OpenBSD (#58)
* make Makefile more portable

tbamud doesn't build on FreeBSD or OpenBSD.  This series of commits are meant to make those builds work by making the Makefile more portable.
I should mention that I've never used tbamud before today, so please read my pull request carefully.

This commit itself fixes the first issue encountered.  I assume wildcard and patsubst are gmake-only.

* Update Makefile.in

It turns out using the backslash to escape newlines in this situation is not portable.

* Delete Makefile.in

Will re-add.  There are problems with this file.  It's marked executable when it shouldn't be, and has ^M characters which prevent builds on OpenBSD.

* re-add Makefile.in